fix json codec
